refactor(service_client): extract _BaseClient with shared helpers
Move pure helpers (URL building, source description, options resolution,
result assembly, retry/error decisions) and shared config into a new
_BaseClient class. DoclingServiceClient now inherits from it.

No behavior change. The async client added in the next commit will share
the same base, avoiding ~500 lines of duplicated logic.
